Fix : Bugs and Divide Route to login
