+++ src/routes/sessionRoutes.js
@@ -1,0 +1,9 @@
+const express = require('express')
+const router = express.Router()
+
+const { sessionCheck } = require('../controllers/sessionControllers')
+const { checkRequest } = require('../middlewares/checkRequest')
+
+router.post('', checkRequest, sessionCheck)
+
+module.exports = router
